use new builder image approach
this commit updates builder images and adds a make target so that we
copy the static executables right from the container image.
it's fast and we don't have to pull it from the internet again and again
if we have the builder image present locally.

This commit also updates go to 1.21.8 patch version.

Signed-off-by: kranurag7 <anurag.kumar@syself.com>