remove schedule link checker workflow
We already check broken links in every PR so we don't need to
have this separate workflow that checks links at a particular time via
cron.

Other reason is that, we don't get notification when this workflow
fails. I think we can remove this because I think this is an over
allocation of resources.

We already have this enabled via `make verify`

Signed-off-by: kranurag7 <anurag.kumar@syself.com>