replace ARG with ENV in builder Dockerfile
I think the versions are not getting updated which are backec by ARG.
This commit replaces the same with ENV so that we get automatic updates
for all the components.

kubectl is excluded because maybe that will be very fast upgrades
multiple times a month and kubectl is pretty stable I think and we
probably don't need really fast upgrades there.

Signed-off-by: kranurag7 <anurag.kumar@syself.com>