add buildx in renovate workflow
renovate workflow failed because we build image in that workflow and
buildx was missing. buildx is now the default builder of docker and we
cannot build images without that.
This commit adds buildx to the renovate workflow so that when renovate
workflow is run and inside the workflow, if we build docker image then
that shouldn't give us any error.

Signed-off-by: kranurag7 <anurag.kumar@syself.com>